Upload relocation extractor/applier script
This script can scan a DRM file, analyze every chunk of it (which can contain an overlay module or resource) to search for relocation data in RedirectList or RELMOD format, and then extract it. It also extracts the raw overlay module, and produces a patched version of it with the relocations applied to it so that it's easier to disassemble.